+++ packages/mobile/src/i18n/locales/ru/errors.ts
@@ -1,6 +1,7 @@
 const errors = {
   auth: 'Ошибка аутентификации ({{ id }})',
   default: 'Что-то пошло не так ({{ id }})',
+  fetch: 'Ошибка соединения с интернетом',
 };
 
 export default errors;