+++ services/backend/src/features/gauges/mutations/index.ts
@@ -1,10 +1,12 @@
 import removeGauge from './removeGauge';
+import removeGauges from './removeGauges';
 import toggleAllGauges from './toggleAllGauges';
 import toggleGauge from './toggleGauge';
 import upsertGauge from './upsertGauge';
 
 export default {
   removeGauge,
+  removeGauges,
   toggleGauge,
   toggleAllGauges,
   upsertGauge,