+++ services/web/src/features/sources/details/toggleAllGauges.mutation.ts
@@ -1,8 +1,16 @@
 import gql from 'graphql-tag';
 
 const TOGGLE_ALL_GAUGES = gql`
-  mutation toggleAllGauges($sourceId: ID!, $enabled: Boolean!) {
-    toggleAllGauges(sourceId: $sourceId, enabled: $enabled) {
+  mutation toggleAllGauges(
+    $sourceId: ID!
+    $enabled: Boolean!
+    $linkedOnly: Boolean
+  ) {
+    toggleAllGauges(
+      sourceId: $sourceId
+      enabled: $enabled
+      linkedOnly: $linkedOnly
+    ) {
       id
       enabled
     }