+++ packages/mobile/CHANGELOG.md
@@ -1,3 +1,26 @@
+## [1.7.66](https://github.com/doomsower/whitewater/compare/@whitewater-guide/mobile@1.7.65...@whitewater-guide/mobile@1.7.66) (2019-05-22)
+
+## [1.7.65](https://github.com/doomsower/whitewater/compare/@whitewater-guide/mobile@1.7.64...@whitewater-guide/mobile@1.7.65) (2019-05-13)
+
+### Bug Fixes
+
+- **mobile:** track extra on exception ([6778354](https://github.com/doomsower/whitewater/commit/6778354))
+
+### Features
+
+- **mobile:** add screen breadcrumbs in sentry ([a981fe1](https://github.com/doomsower/whitewater/commit/a981fe1))
+- **mobile:** do not track failed network requests when offline ([edcce3b](https://github.com/doomsower/whitewater/commit/edcce3b))
+
+## [1.7.64](https://github.com/doomsower/whitewater/compare/@whitewater-guide/mobile@1.7.63...@whitewater-guide/mobile@1.7.64) (2019-05-13)
+
+### Bug Fixes
+
+- **mobile:** improve chips ui ([ffbea6c](https://github.com/doomsower/whitewater/commit/ffbea6c))
+
+### Features
+
+- **clients:** add fetch type to AppError ([df36d05](https://github.com/doomsower/whitewater/commit/df36d05))
+
 ## [1.7.63](https://github.com/doomsower/whitewater/compare/@whitewater-guide/mobile@1.7.62...@whitewater-guide/mobile@1.7.63) (2019-04-28)
 
 ### Bug Fixes