+++ services/boompromo/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.144](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.143...@whitewater-guide/boompromo@0.0.144) (2019-05-22)
+
+## [0.0.143](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.142...@whitewater-guide/boompromo@0.0.143) (2019-05-13)
+
 ## [0.0.142](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.141...@whitewater-guide/boompromo@0.0.142) (2019-04-27)
 
 ## [0.0.141](https://github.com/doomsower/whitewater/compare/@whitewater-guide/boompromo@0.0.140...@whitewater-guide/boompromo@0.0.141) (2019-04-26)